refactor(plugin): split FlipSmartApiClient into transport + endpoint groups
Part of #731 (epic #728).

- New com.flipsmart.api.ApiHttpTransport owns OkHttp/Gson/config, ALL
  auth/session state (jwt, refresh token, premium/rsn flags, authLock,
  pendingAuthFuture, callbacks), the execute*/auth-coalescing/401-retry
  machinery, and the auth + device-auth endpoint methods.
- New com.flipsmart.api.endpoints.* groups (Flips, Transaction, ActiveFlip,
  OfferAction, MarketData, BankSnapshot, Blocklist, Webhook, Motd,
  TradeStation) delegate HTTP to the transport and own only their response
  caches (analysis, wiki price, daily volume).
- FlipSmartApiClient stays a @Singleton facade preserving every public
  signature via one-line delegations (2464 -> 526 lines); no consumer or
  test outside the api packages changes. Behavior identical (URLs, JSON
  keys, cache TTLs, retry, threading unchanged).
